dcp-815 move config to relevant packages
